작업파일 정리 sql 전부 업로드

+++ 2장_테이블 제약조건.sql
@@ -77,7 +77,7 @@ INSERT INTO `User4`(`gender`, `addr`) VALUES(1, '부산시');
 CREATE TABLE `User5` (
 			`name`		VARCHAR(10) NOT NULL,
 			`gender`		TINYINT,
-			`age`			INT DEFAULT 1,
+			`age`			TINYINT DEFAULT 1,
 			`addr`		VARCHAR(10)
 			);
